% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-reinterpret.R
\name{as.data.frame.PeakReinterpretation}
\alias{as.data.frame.PeakReinterpretation}
\title{Flatten champions from a PeakReinterpretation}
\usage{
\method{as.data.frame}{PeakReinterpretation}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{PeakReinterpretation}.}

\item{row.names, optional}{Unused; required for \code{as.data.frame} S3.}

\item{...}{Unused.}
}
\value{
A data.frame.
}
\description{
Flatten champions from a PeakReinterpretation
}
