% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{assignPheno}
\alias{assignPheno}
\title{Assign phenotype data to a LazyGas object}
\usage{
assignPheno(object, pheno, rename = NULL, ...)
}
\arguments{
\item{object}{A LazyGas object}

\item{pheno}{A data.frame of phenotype data that must contain a sample id
column named \code{id} or \code{ID}}
}
\description{
Assign phenotype data to a LazyGas object
}
