% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{assignPvalues}
\alias{assignPvalues}
\title{Assign phenotype data to a LazyGas object}
\usage{
assignPvalues(
  object,
  pheno_name,
  p_values,
  geno_format = c("genotype", "dosage", "haplotype"),
  conv_fun = NULL,
  formula = "phe ~ add",
  null_formula = NULL,
  kruskal = NULL,
  ...
)
}
\arguments{
\item{object}{A LazyGas object.}

\item{pheno_name}{A character vector.}

\item{p_values}{A numeric matrix or data.frame with the number of rows matching the number of markers.}
}
\description{
Assign phenotype data to a LazyGas object
}
