% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{buildGWASscan}
\alias{buildGWASscan}
\title{Build a lazyGWAS object}
\usage{
buildGWASscan(gds, pheno, pvalues_fn, sampleID_pheno = NULL)
}
\arguments{
\item{gds}{A gbsrGenotypeData object}

\item{pheno}{Phenotype data.frame}

\item{pvalues_fn}{pvalues file names}

\item{sampleID_pheno}{Sample ID in phenotype data}
}
\description{
Build a lazyGWAS object
}
