% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{buildLazyGWAS}
\alias{buildLazyGWAS}
\title{Build a lazyGWAS object}
\usage{
buildLazyGWAS(gds, pheno, sampleID_pheno = NULL)
}
\arguments{
\item{gds}{A gbsrGenotypeData object}

\item{pheno}{Phenotype data.frame}

\item{sampleID_pheno}{Sample ID in phenotype data}
}
\description{
Build a lazyGWAS object
}
