% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{buildQTLscan}
\alias{buildQTLscan}
\title{Build QTLscan object}
\usage{
buildQTLscan(
  geno,
  pheno,
  pvalues_fn,
  sampleID_geno = NULL,
  sampleID_pheno = NULL,
  marker_chr,
  marker_pos,
  geno_fmt = "dosage",
  geno_levels = 0:2
)
}
\arguments{
\item{geno}{Genotype data.frame}

\item{pheno}{Phenotype data.frame}

\item{pvalues_fn}{pvalues file names}

\item{sampleID_geno}{Sample ID in genotype data}

\item{sampleID_pheno}{Sample ID in phenotype data}

\item{marker_chr}{Chromosome ID of markers}

\item{marker_pos}{Physical positin of markers}

\item{geno_fmt}{Genotype data format}

\item{geno_levels}{Factor levels of genotype data}
}
\description{
Build QTLscan object
}
