% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calcGenomicInflation}
\alias{calcGenomicInflation}
\title{Genomic inflation factor (lambda GC)}
\usage{
calcGenomicInflation(p_values)
}
\arguments{
\item{p_values}{Numeric vector of p-values (NA omitted).}
}
\description{
Genomic inflation factor (lambda GC)
}
