% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-reinterpret.R
\name{champion}
\alias{champion}
\title{Champion Gene_ID helper}
\usage{
champion(x, peak, view)
}
\arguments{
\item{x}{PeakReinterpretation}

\item{peak}{Peak id}

\item{view}{View name}
}
\value{
Character scalar or \code{NA_character_}
}
\description{
Champion Gene_ID helper
}
