% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-reinterpret.R
\name{conflictTable}
\alias{conflictTable}
\title{Conflict table from a PeakReinterpretation}
\usage{
conflictTable(x)
}
\arguments{
\item{x}{A \code{PeakReinterpretation}.}
}
\value{
A data.frame of conflicts across peaks.
}
\description{
Conflict table from a PeakReinterpretation
}
