% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-evidence.R
\name{.floor_score_vec}
\alias{.floor_score_vec}
\title{Floor normalized evidence scores (default minimum 0.1)}
\usage{
.floor_score_vec(x, floor = 0.1)
}
\description{
Floor normalized evidence scores (default minimum 0.1)
}
\keyword{internal}
