% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm-report-e0.R
\name{.gff_gene_protein_map}
\alias{.gff_gene_protein_map}
\title{Gene_ID ↔ transcript / protein IDs from GFF attributes}
\usage{
.gff_gene_protein_map(gff)
}
\description{
Gene_ID ↔ transcript / protein IDs from GFF attributes
}
\keyword{internal}
