% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakcall.R
\name{.ld_identity_categorical}
\alias{.ld_identity_categorical}
\title{Identity / match rate LD for categorical (e.g. haplotype) genotypes.}
\usage{
.ld_identity_categorical(geno, peak_variant_idx, n_threads = NULL)
}
\description{
Identity / match rate LD for categorical (e.g. haplotype) genotypes.
}
\keyword{internal}
