% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-server.R
\name{.ollama_cli_path}
\alias{.ollama_cli_path}
\title{Locate the Ollama command-line binary}
\usage{
.ollama_cli_path()
}
\value{
Character scalar (empty if not found).
}
\description{
Locate the Ollama command-line binary
}
\keyword{internal}
