% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.R
\name{.peak_block_markers}
\alias{.peak_block_markers}
\title{Peak-block markers for one peak (recalc, else peakcall)}
\usage{
.peak_block_markers(object, pheno_name, peak_id)
}
\description{
Peak-block markers for one peak (recalc, else peakcall)
}
\keyword{internal}
