% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase1-prep.R
\name{.phase1_filter_missing_maf_from_info}
\alias{.phase1_filter_missing_maf_from_info}
\title{Filter markers by missing rate and MAF using INFO AN / AF.}
\usage{
.phase1_filter_missing_maf_from_info(object, missing = 0.2, maf = 0.05)
}
\arguments{
\item{object}{GbsrGenotypeData / LazyGas}

\item{missing}{Max missing genotype rate (default 0.2)}

\item{maf}{Min minor allele frequency (default 0.05)}
}
\description{
Filter markers by missing rate and MAF using INFO AN / AF.
}
\keyword{internal}
