% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-query-vocab.R
\name{.phenotype_linked_phrase_strings}
\alias{.phenotype_linked_phrase_strings}
\title{Linked phrase strings from a PhenotypeQuery field}
\usage{
.phenotype_linked_phrase_strings(x)
}
\description{
Linked phrase strings from a PhenotypeQuery field
}
\keyword{internal}
