% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-query.R
\name{.phenotype_query_as_search_text}
\alias{.phenotype_query_as_search_text}
\title{Convert a PhenotypeQuery to a single search string}
\usage{
.phenotype_query_as_search_text(query)
}
\arguments{
\item{query}{A \code{PhenotypeQuery} or character string.}
}
\description{
Convert a PhenotypeQuery to a single search string
}
\keyword{internal}
