% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm-report-b2.R
\name{.report_basic_info_html}
\alias{.report_basic_info_html}
\title{Basic information as HTML}
\usage{
.report_basic_info_html(object, pheno_name, language = "en")
}
\description{
Basic information as HTML
}
\keyword{internal}
