% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.store_read_simple_candidate}
\alias{.store_read_simple_candidate}
\title{Read E0 simple candidate table}
\usage{
.store_read_simple_candidate(object, pheno_name)
}
\description{
Read E0 simple candidate table
}
\keyword{internal}
