% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.store_write_gene_protein_map}
\alias{.store_write_gene_protein_map}
\title{Write GFF Gene↔transcript/protein map}
\usage{
.store_write_gene_protein_map(object, protein_map)
}
\description{
Write GFF Gene↔transcript/protein map
}
\keyword{internal}
