% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-rank.R
\name{.store_write_phenotype_rank}
\alias{.store_write_phenotype_rank}
\title{Save a phenotype-rank table to the companion store}
\usage{
.store_write_phenotype_rank(object, rank_result, pheno_name, query_id)
}
\description{
Save a phenotype-rank table to the companion store
}
\keyword{internal}
