% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{.store_write_simple_candidate}
\alias{.store_write_simple_candidate}
\title{Write E0 simple candidate Parquet (Gene_ID + position)}
\usage{
.store_write_simple_candidate(object, pheno_name, simple)
}
\description{
Write E0 simple candidate Parquet (Gene_ID + position)
}
\keyword{internal}
