% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{importLazyGasResults}
\alias{importLazyGasResults}
\title{Import lazyGas results from the legacy GDS \verb{lazygas/} subtree into the companion store.}
\usage{
importLazyGasResults(object)
}
\arguments{
\item{object}{A LazyGas object using Parquet or SQLite storage.}
}
\description{
Import lazyGas results from the legacy GDS \verb{lazygas/} subtree into the companion store.
}
