% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\docType{class}
\name{LazyGas-class}
\alias{LazyGas-class}
\alias{LazyGas}
\title{Class \code{LazyGas}}
\description{
The \code{LazyGas} class is the main class of \code{lazyGas} and
user work with this class object.
}
\details{
The \code{LazyGas} class is an extention of \code{GbsrGenotypeData} in the
\link[GBScleanR:GBScleanR]{GBScleanR} package.
}
\section{Slots}{

\describe{
\item{\code{lazydata}}{A \link{list} object.}
}}

\examples{
# `buildLazyGas()` initialize the `LazyGas` object.

# Load a GDS file and instantiate a `LazyGas` object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
lgas <- buildLazyGas(gds_fn)

# Close connection to the GDS file.
closeGDS(lgas)

}
