% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-server.R
\name{listOllamaModels}
\alias{listOllamaModels}
\title{List models available on a local Ollama server}
\usage{
listOllamaModels(base_url = NULL)
}
\arguments{
\item{base_url}{Ollama base URL.}
}
\value{
Character vector of model names.
}
\description{
List models available on a local Ollama server
}
