% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.R
\name{makeCanditeList}
\alias{makeCanditeList}
\title{Generate interactive table for a candidate gene list}
\usage{
makeCanditeList(object, pheno, out_fn, peak_id = NULL)
}
\arguments{
\item{object}{QTLscan object}

\item{out_fn}{Prefix of output file}

\item{pehno}{Phenotype names to be drawn}
}
\description{
Generate interactive table for a candidate gene list
}
