% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.R
\name{plotCredibleSet}
\alias{plotCredibleSet}
\title{Plot credible-set PIP values for a peak}
\usage{
plotCredibleSet(object, pheno = NULL, peak_id = NULL, ...)
}
\arguments{
\item{object}{A \code{LazyGas} object or a data.frame from \code{\link[=calcCredibleSet]{calcCredibleSet()}}.}

\item{pheno, peak_id}{Required when \code{object} is a \code{LazyGas} object.}

\item{...}{Passed to \code{\link[=calcCredibleSet]{calcCredibleSet()}} when needed.}
}
\description{
Plot credible-set PIP values for a peak
}
