% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Draw a Manhattan plot}
\usage{
plotManhattan(
  object,
  pheno = NULL,
  chr = NULL,
  start = NULL,
  end = NULL,
  signif = 0.05,
  out_fn = "",
  ...
)
}
\arguments{
\item{object}{QTLscan object}

\item{chr}{Chromosome ID to be drawn}

\item{start}{Start position of the range to be drawn}

\item{end}{End position of the range to be drawn}

\item{signif}{Expression to define the significant markers}

\item{out_fn}{Prefix of output file}

\item{pehno}{Phenotype names to be drawn}
}
\description{
Draw a Manhattan plot
}
