% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog.R
\name{plotOrthologSummary}
\alias{plotOrthologSummary}
\title{Plot ortholog overlap among annotated candidates}
\usage{
plotOrthologSummary(annotated)
}
\arguments{
\item{annotated}{Output of \code{\link[=annotateOrthologs]{annotateOrthologs()}}.}
}
\description{
Plot ortholog overlap among annotated candidates
}
