% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{plotPeaks}
\alias{plotPeaks}
\title{Plots the peaks for the specified phenotype and chromosome region.}
\usage{
plotPeaks(
  object,
  pheno = NULL,
  chr = NULL,
  start = NULL,
  end = NULL,
  recalc = TRUE,
  ...
)
}
\arguments{
\item{object}{A LazyGas object.}

\item{pheno}{The phenotype(s) to plot. Can be NULL, numeric, logical, or character.}

\item{chr}{The chromosome to plot. Can be NULL.}

\item{start}{The start position on the chromosome. Can be NULL.}

\item{end}{The end position on the chromosome. Can be NULL.}

\item{recalc}{Logical. If TRUE, use recalculated scan data.}

\item{...}{Additional arguments.}
}
\value{
Generates plots for the specified peaks.
}
\description{
Plots the peaks for the specified phenotype and chromosome region.
}
