% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQQ}
\alias{plotQQ}
\title{QQ plot for GWAS p-values}
\usage{
plotQQ(object, pheno = NULL, max_points = 100000L)
}
\arguments{
\item{object}{A \code{LazyGas} object, or a numeric p-value vector.}

\item{pheno}{Phenotype name or index when \code{object} is \code{LazyGas}.}

\item{max_points}{Maximum points drawn (full data still used for
\eqn{\lambda_{GC}}). Large genome-wide scans are thinned for plotting.}
}
\description{
QQ plot for GWAS p-values
}
