% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype-reinterpret.R
\name{reinterpretDefaultViews}
\alias{reinterpretDefaultViews}
\title{Default reinterpretation views}
\usage{
reinterpretDefaultViews()
}
\value{
Character vector of view names.
}
\description{
Default reinterpretation views
}
