% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{runLazyGasShiny}
\alias{runLazyGasShiny}
\title{Launch the lazyGas Shiny explorer}
\usage{
runLazyGasShiny(...)
}
\arguments{
\item{...}{Arguments passed to \code{shiny::runApp()}.}
}
\description{
Launch the lazyGas Shiny explorer
}
