% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{scanGWAS}
\alias{scanGWAS}
\alias{scanGWAS.lazyGWAS}
\title{Scan GWAS}
\usage{
scanGWAS(x, ...)

\method{scanGWAS}{lazyGWAS}(x, out_fn, nPC = 2, n_core = 1)
}
\arguments{
\item{x}{A lazyGWAS object}

\item{out_fn}{Prefix of output file name}

\item{nPC}{The number of principal components in the regression model}

\item{n_core}{The number of threads for parallel computing}
}
\description{
Scan GWAS
}
