% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{scanQTL}
\alias{scanQTL}
\alias{scanQTL.lazyQTL}
\title{Scan QTL}
\usage{
scanQTL(x, ...)

\method{scanQTL}{lazyQTL}(x, out_fn, formula = "", makeDF_FUN = NULL)
}
\arguments{
\item{x}{A lazyQTL object}

\item{out_fn}{Prefix of output file name}

\item{formula}{The formula of the regression model}

\item{makeDF_FUN}{The function to modify model data for the regression}
}
\description{
Scan QTL
}
