% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{summarizeCrossTraitPeaks}
\alias{summarizeCrossTraitPeaks}
\title{Summarize cross-trait peak clusters}
\usage{
summarizeCrossTraitPeaks(object, recalc = TRUE)
}
\arguments{
\item{object}{A \code{LazyGas} object, or a cluster table from
\code{\link[=clusterCrossTraitPeaks]{clusterCrossTraitPeaks()}}.}

\item{recalc}{Use recalculated peaks when collecting peaks from \code{object}.}
}
\description{
Summarize cross-trait peak clusters
}
