% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{summarizeGWASQC}
\alias{summarizeGWASQC}
\title{Summarize GWAS quality metrics}
\usage{
summarizeGWASQC(object, pheno = NULL, small_n_threshold = 30L, store = TRUE)
}
\arguments{
\item{object}{A \code{LazyGas} object.}

\item{pheno}{Phenotype name or index; if \code{NULL}, all phenotypes.}

\item{small_n_threshold}{Sample size below which a warning flag is set.}

\item{store}{If \code{TRUE}, write results to the companion store.}
}
\description{
Summarize GWAS quality metrics
}
