% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog.R
\name{summarizeOrthologMatches}
\alias{summarizeOrthologMatches}
\title{Summarize ortholog conservation across candidate lists}
\usage{
summarizeOrthologMatches(annotated)
}
\arguments{
\item{annotated}{Output of \code{\link[=annotateOrthologs]{annotateOrthologs()}}.}
}
\description{
Summarize ortholog conservation across candidate lists
}
