% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tenor.R
\name{tenorDatasetCatalog}
\alias{tenorDatasetCatalog}
\title{Built-in TENOR dataset catalog used by RAP-DB Expression (TENOR) UI}
\usage{
tenorDatasetCatalog()
}
\value{
A data.frame with columns \code{dataset} and \code{publication}.
}
\description{
Built-in TENOR dataset catalog used by RAP-DB Expression (TENOR) UI
}
