% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{buildLazyGas}
\alias{buildLazyGas}
\title{Build a LazyGas object}
\usage{
buildLazyGas(gds_fn = "", load_filter = TRUE, overwrite = FALSE)
}
\arguments{
\item{gds_fn}{A path to a GDS file}

\item{load_filter}{A logical value to indicate whether apply filering stored in the input GDS file to samples and markers.}
}
\description{
Build a LazyGas object
}
