% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{buildLazyGas}
\alias{buildLazyGas}
\title{Build a LazyGas object}
\usage{
buildLazyGas(
  gds_fn = "",
  load_filter = TRUE,
  overwrite = FALSE,
  create_gds = NULL
)
}
\arguments{
\item{gds_fn}{A path to a GDS file}

\item{load_filter}{A logical value to indicate whether apply filrering stored in the input GDS file to samples and markers.}

\item{create_gds}{A named list to create a new GDS file with specified genotype and marker information. See the Details section.}
}
\description{
Build a LazyGas object
}
\details{
As the default, the \code{buildLazyGas()} function loads the genotype and marker
information from a GDS file specified to the \code{gds_fn} argument.
If a named list storing genotype and marker information was specified to the
\code{create_gds} argument, a GDS file will be created with supplied genotype and
marker information. the \code{create_gds} list should have the following elements.
\itemize{
\item genotype:
\item sample.id:
\item snp.id:
\item snp.rs.id:
\item snp.chromosome:
\item snp.position:
\item snp.allele:
\item haplotype:
\item dosage:
Either genotype, haplotype, or dosage should
be supplied in the \code{create_gds} list. If \code{sample.id}, \code{snp.id}, and
\code{snp.rs.id} were left as \code{NULL}, serial numbers will be assigned as sample
and SNP IDs. Random alleles will be assigned if \code{snp.allele = NULL}.
\code{snp.chromosome} and \code{snp.marker} must be supplied.
}
}
