% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{callPeakBlock}
\alias{callPeakBlock}
\title{Call peack blocks}
\usage{
callPeakBlock(object, signif = 0.05, threshold = 0.8, n_threads = NULL, ...)
}
\arguments{
\item{object}{QTLscan object}

\item{signif}{expression to define significant markers}

\item{rsquare}{threshold on squared R values to define peak blocks}
}
\description{
Call peack blocks
}
