% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-evidence.R
\name{.finemap_pip_lookup}
\alias{.finemap_pip_lookup}
\title{Max PIP per gene from 95\% credible set ∩ SnpEff Gene_ID}
\usage{
.finemap_pip_lookup(object, pheno_name, candidate)
}
\description{
Max PIP per gene from 95\% credible set ∩ SnpEff Gene_ID
}
\keyword{internal}
