% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_candidate_genes.R
\name{.keyword_terms_from_query}
\alias{.keyword_terms_from_query}
\title{Build content keyword / phrase list from a phenotype query}
\usage{
.keyword_terms_from_query(query)
}
\description{
Build content keyword / phrase list from a phenotype query
}
\keyword{internal}
