% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{getGenoPerMarker}
\alias{getGenoPerMarker}
\title{Confirm data structure of genotype data per marker}
\usage{
getGenoPerMarker(object, geno_format, marker_id, ...)
}
\description{
Confirm data structure of genotype data per marker
}
