% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{getPheno}
\alias{getPheno}
\alias{getPheno,LazyGas-method}
\title{Get Phenotype Data from LazyGas Object}
\usage{
getPheno(object, ...)

\S4method{getPheno}{LazyGas}(object)
}
\arguments{
\item{object}{An object of class \code{LazyGas}.}
}
\value{
A list containing phenotype data from the \code{LazyGas} object.
}
\description{
Get Phenotype Data from LazyGas Object
}
