% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{haploPlot}
\alias{haploPlot}
\alias{haploPlot,LazyGas-method}
\title{Generate Haplotype Plots}
\usage{
haploPlot(object, pheno, recalc, ...)

\S4method{haploPlot}{LazyGas}(object, pheno, recalc)
}
\arguments{
\item{object}{A LazyGas object.}

\item{pheno}{The phenotype to plot.}

\item{recalc}{Logical. If TRUE, use recalculated scan data.}

\item{...}{Additional arguments.}
}
\value{
A list of ggplot objects representing the haplotype plots.
}
\description{
This function generates haplotype plots for the specified phenotype.
Plot titles include peak location, \code{-log10P}, and the additive scan
effect direction when \code{Coef.add} is available in stored scan results.
}
