% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{haploPlot}
\alias{haploPlot}
\alias{haploPlot.peakCall}
\alias{haploPlot.data.frame}
\title{Haplotype analysis for peak blocks}
\usage{
haploPlot(x, ...)

\method{haploPlot}{peakCall}(x, out_fn, out_fmt)

\method{haploPlot}{data.frame}(x, geno, pheno, out_fn, out_fmt)
}
\arguments{
\item{x}{input object}

\item{out_fn}{prefix of output file name}

\item{geno}{genotype data.frame}

\item{pheno}{phenotype data.frame}

\item{peakblock}{peak block data.frame}
}
\description{
Haplotype analysis for peak blocks
}
