% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazygas-class.R
\docType{class}
\name{LazyGas-class}
\alias{LazyGas-class}
\alias{LazyGas}
\title{The \code{LazyGas} class is an extention of \code{GbsrGenotypeData} in the
\link[GBScleanR:GBScleanR]{GBScleanR} package.}
\description{
The \code{LazyGas} class is an extention of \code{GbsrGenotypeData} in the
\link[GBScleanR:GBScleanR]{GBScleanR} package.
}
\section{Slots}{

\describe{
\item{\code{lazydata}}{A \link{list} object (phenotype data).}

\item{\code{store}}{Companion storage configuration (\code{mode}, \code{path}, \code{gds_fn}).}
}}

\examples{
# `buildLazyGas()` initialize the `LazyGas` object.

# Load a GDS file and instantiate a `LazyGas` object.
gds_fn <- system.file("extdata", "sample.gds", package = "GBScleanR")
lgas <- buildLazyGas(gds_fn)

# Close connection to the GDS file.
closeGDS(lgas)

}
