% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{listCandidate}
\alias{listCandidate}
\alias{listCandidate.peakCall}
\alias{listCandidate.data.frame}
\title{List up candidate genes}
\usage{
listCandidate(x, ...)

\method{listCandidate}{peakCall}(x, annotation_fn, gff_fn, snpeff_fn = NULL, out_fn)

\method{listCandidate}{data.frame}(x, annotation_fn, gff_fn, snpeff_fn = NULL, out_fn, scan = "QTL")
}
\arguments{
\item{x}{peakcall data.frame}

\item{annotation_fn}{path to an annotation information file}

\item{gff_fn}{path to a GFF file}

\item{snpeff_fn}{path to a snpEff file}

\item{out_fn}{prefix of output file name}

\item{scan}{scan type}
}
\description{
List up candidate genes
}
