% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{makeConvFun}
\alias{makeConvFun}
\title{Make a conversion function}
\usage{
makeConvFun(
  geno_format = c("genotype", "corrected", "dosage", "haplotype"),
  n_levels = 3,
  as_categorical = FALSE
)
}
\description{
Make a conversion function
}
