% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{makeInteractiveSummary}
\alias{makeInteractiveSummary}
\title{Generate interactive summary}
\usage{
makeInteractiveSummary(object, pheno, out_fn)
}
\arguments{
\item{object}{QTLscan object}

\item{out_fn}{Prefix of output file}

\item{pehno}{Phenotype names to be drawn}
}
\description{
Generate interactive summary
}
