% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{plotManhattan}
\alias{plotManhattan}
\alias{plotManhattan.QTLscan}
\alias{plotManhattan.GWASscan}
\alias{plotManhattan.character}
\alias{plotManhattan.data.frame}
\title{Draw a Manhattan plot}
\usage{
plotManhattan(x, ...)

\method{plotManhattan}{QTLscan}(
  x,
  pheno = NULL,
  chr = NULL,
  start = NULL,
  end = NULL,
  signif = NULL,
  out_fn = "",
  out_fmt = "pdf"
)

\method{plotManhattan}{GWASscan}(
  x,
  pheno = NULL,
  chr = NULL,
  start = NULL,
  end = NULL,
  signif = NULL,
  out_fn = "",
  out_fmt = "pdf"
)

\method{plotManhattan}{character}(
  x,
  chr = NULL,
  start = NULL,
  end = NULL,
  signif = NULL,
  out_fn = "",
  out_fmt = "pdf"
)

\method{plotManhattan}{data.frame}(
  x,
  chr = NULL,
  start = NULL,
  end = NULL,
  signif = NULL,
  out_fn = "",
  out_fmt = "pdf"
)
}
\arguments{
\item{x}{data.frame of pvalue data}

\item{chr}{Chromosome ID to be drawn}

\item{start}{Start position of the range to be drawn}

\item{end}{End position of the range to be drawn}

\item{signif}{Expression to define the significant markers}

\item{out_fn}{Prefix of output file}

\item{out_fmt}{Output image format}

\item{pehno}{Phenotype names to be drawn}
}
\description{
Draw a Manhattan plot
}
