% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrait.R
\name{plotMultiTraitOverview}
\alias{plotMultiTraitOverview}
\title{Plot multi-trait peak overview}
\usage{
plotMultiTraitOverview(
  object,
  recalc = TRUE,
  dist_threshold = 500000L,
  r2_threshold = 0.6
)
}
\arguments{
\item{object}{A \code{LazyGas} object.}

\item{recalc}{Use recalculated peaks when clustering.}

\item{dist_threshold, r2_threshold}{Passed to \code{\link[=clusterCrossTraitPeaks]{clusterCrossTraitPeaks()}}.}
}
\description{
Plot multi-trait peak overview
}
