% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\name{plotPheno}
\alias{plotPheno}
\title{Assign phenotype data to the samples in the LazyGas object}
\usage{
plotPheno(
  object,
  pheno = 1,
  xlab = "phenotype",
  axis_title_size = 14,
  axis_text_size = 12,
  fill = "skyblue",
  color = "darkblue",
  boxplot = TRUE,
  ...
)
}
\arguments{
\item{object}{A LazyGas object}

\item{pheno}{A data.frame of phenotype data that must contain a sample id
column named \code{id} or \code{ID}}
}
\description{
Assign phenotype data to the samples in the LazyGas object
}
