% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{recalcAssoc}
\alias{recalcAssoc}
\alias{recalcAssoc,LazyGas-method}
\title{Recalculate Associations}
\usage{
recalcAssoc(
  object,
  n_threads = NULL,
  refine_position = FALSE,
  grouping_threshold = 0.05,
  ...
)

\S4method{recalcAssoc}{LazyGas}(object, n_threads, refine_position, grouping_threshold)
}
\arguments{
\item{object}{A LazyGas object.}

\item{n_threads}{description}

\item{refine_position}{}
}
\value{
None. The function modifies the LazyGas object in place.
}
\description{
Recalculate associations using the LazyGas object and specified r-square threshold.
}
