% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{recalcAssoc}
\alias{recalcAssoc}
\alias{recalcAssoc,LazyGas-method}
\title{Recalculate Associations}
\usage{
recalcAssoc(object, signif = 0.05, threshold = 0.6, ...)

\S4method{recalcAssoc}{LazyGas}(object, signif, threshold)
}
\arguments{
\item{object}{A LazyGas object.}
}
\value{
None. The function modifies the LazyGas object in place.
}
\description{
Recalculate associations using the LazyGas object and specified r-square threshold.
}
