% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scanAssoc}
\alias{scanAssoc}
\title{Scan QTL}
\usage{
scanAssoc(
  object,
  formula = "phe ~ add",
  null_formula = NULL,
  conv_fun = NULL,
  fixed_effect = NULL,
  geno_format = c("genotype", "corrected", "dosage", "haplotype"),
  kruskal = NULL,
  method = c("glm", "mlm"),
  ...
)
}
\arguments{
\item{object}{A LazyGas object}

\item{formula}{The formula of the regression model}

\item{conv_fun}{The function to convert genotype data to a model matrix for the regression}

\item{fixed_effect}{A data.frame of fixed effect(s) incorporated in the regression model. The column names must match the terms specified in \code{formula} and \code{null_formula.} The number of rows must match the number of samples.}

\item{out_fn}{Prefix of output file name}
}
\description{
Scan QTL
}
