% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_functions.R
\name{scanAssoc}
\alias{scanAssoc}
\title{Scan QTL}
\usage{
scanAssoc(
  object,
  formula,
  makeDF_FUN = NULL,
  geno_format = c("genotype", "dosage", "haplotype"),
  kruskal = NULL,
  ...
)
}
\arguments{
\item{object}{A LazyGas object}

\item{formula}{The formula of the regression model}

\item{makeDF_FUN}{The function to modify model data for the regression}

\item{out_fn}{Prefix of output file name}
}
\description{
Scan QTL
}
