% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{scanAssoc}
\alias{scanAssoc}
\title{Scan QTL}
\usage{
scanAssoc(
  object,
  formula = "phe ~ add",
  conv_fun = NULL,
  geno_format = c("genotype", "dosage", "haplotype"),
  kruskal = NULL,
  method = c("glm", "mlm"),
  ...
)
}
\arguments{
\item{object}{A LazyGas object}

\item{formula}{The formula of the regression model}

\item{conv_fun}{The function to convert genotype data to a model matrix for the regression}

\item{out_fn}{Prefix of output file name}
}
\description{
Scan QTL
}
